Implementando produto no carrinho de compra
