--- src/produto.java
@@ -1,83 +1,0 @@
-
-import java.util.Date;
-
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-/**
- *
- * @author jardel
- */
-public class produto {
-    
-    private String nome;
-    private String unidade;
-    private double preco;
-    private Date validade;
-    
-    public produto(){
-    
-}
-
-    /**
-     * @return the nome
-     */
-    public String getNome() {
-        return nome;
-    }
-
-    /**
-     * @param nome the nome to set
-     */
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    /**
-     * @return the unidade
-     */
-    public String getUnidade() {
-        return unidade;
-    }
-
-    /**
-     * @param unidade the unidade to set
-     */
-    public void setUnidade(String unidade) {
-        this.unidade = unidade;
-    }
-
-    /**
-     * @return the preco
-     */
-    public double getPreco() {
-        return preco;
-    }
-
-    /**
-     * @param preco the preco to set
-     */
-    public void setPreco(double preco) {
-        this.preco = preco;
-    }
-
-    /**
-     * @return the validade
-     */
-    public Date getValidade() {
-        return validade;
-    }
-
-    /**
-     * @param validade the validade to set
-     */
-    public void setValidade(Date validade) {
-        this.validade = validade;
-    }
-    
-    
-    
-}
